test/e2e/pyproject.toml
[project]
name = "pptx-renderer-e2e"
version = "0.1.0"
description = "E2E tests for pptx-renderer"
requires-python = ">=3.10"
dependencies = [
  "pytest>=8.0",
  "playwright>=1.40",
  "Pillow>=10.0",
  "scikit-image>=0.22",
  "PyMuPDF>=1.23",
  "lxml>=5.0",
  "numpy>=1.26",
  "opencv-python-headless>=4.9",
  "requests>=2.31",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = []

[tool.pytest.ini_options]
testpaths = ["."]
asyncio_mode = "auto"
timeout = 180
markers = [
  "local_oracle: tests that require local Microsoft PowerPoint automation",
  "local_integration: broader local integration checks",
]
